refactor(runbotics-orchestrator-ui): :recycle: Extracts 2 components from too long form component
