+++ runbotics/runbotics-orchestrator-ui/src/src-app/views/credentials/Credential/CreateGeneralInfo/CreateGeneralInfoFormDialog.tsx
@@ -1,29 +1,23 @@
-/* eslint-disable max-lines-per-function */
 import { FC, useState } from 'react';
 
-import { Grid, TextField, Typography, SelectChangeEvent } from '@mui/material';
+import { Grid, Typography } from '@mui/material';
 
 import { useRouter } from 'next/router';
 import { useSnackbar } from 'notistack';
 
-import { PrivilegeType } from 'runbotics-common';
-
 import CustomDialog from '#src-app/components/CustomDialog';
-import useAuth from '#src-app/hooks/useAuth';
 import useTranslations from '#src-app/hooks/useTranslations';
 
-import { useDispatch, useSelector } from '#src-app/store';
-import { credentialCollectionsSelector } from '#src-app/store/slices/CredentialCollections';
+import { useDispatch } from '#src-app/store';
 import { credentialsActions } from '#src-app/store/slices/Credentials';
 import { Content, Form } from '#src-app/views/utils/FormDialog.styles';
 
-import { CollectionDropdown } from './CollectionDropdown';
-import { TemplateDropdown } from './TemplateDropdown';
+import { GeneralInfoSelectFields } from './GeneralInfoSelectFields';
+import { GeneralInfoTextFields } from './GeneralInfoTextInputs';
 import { CreateCredentialDto } from '../Credential.types';
 import {
     getInitialCredentialData,
-    getInitialFormValidationState,
-    inputErrorMessages
+    getInitialFormValidationState
 } from '../EditCredential/EditCredential.utils';
 
 interface CreateGeneralInfoProps {
@@ -37,13 +31,9 @@ const CreateGeneralInfoFormDialog: FC<CreateGeneralInfoProps> = ({ onClose, open
     const router = useRouter();
     const { enqueueSnackbar } = useSnackbar();
     const collectionId = router.query.collectionId ? (router.query.collectionId as string) : null;
-    const { user: currentUser } = useAuth();
 
     const [credentialFormState, setCredentialFormState] = useState<CreateCredentialDto>(getInitialCredentialData(collectionId));
     const [formValidationState, setFormValidationState] = useState(getInitialFormValidationState(collectionId));
-
-    const { credentialCollections } = useSelector(credentialCollectionsSelector);
-    const availableCredentialCollections = credentialCollections?.filter(collection => collection.credentialCollectionUser.find(collectionUser => collectionUser.userId === currentUser.id && collectionUser.privilegeType === PrivilegeType.WRITE));
 
     const isFormValid = () => Object.values(formValidationState).every(Boolean);
 
@@ -59,6 +49,7 @@ const CreateGeneralInfoFormDialog: FC<CreateGeneralInfoProps> = ({ onClose, open
 
     const handleSubmit = () => {
         if (!isFormValid()) {
+            setFormValidationState(prevValues => ({...prevValues, edited: true}));
             enqueueSnackbar(translate('Credential.Add.ValidationFail.Info'), { variant: 'error' });
             return;
         }
@@ -84,39 +75,6 @@ const CreateGeneralInfoFormDialog: FC<CreateGeneralInfoProps> = ({ onClose, open
             });
     };
 
-    const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const { name, value } = e.target;
-
-        setCredentialFormState(prevState => ({
-            ...prevState,
-            [name]: value
-        }));
-
-        setFormValidationState(prevState => ({
-            ...prevState,
-            [name]: value.trim() !== '',
-            edited: true
-        }));
-    };
-
-    const handleDropdownChange = (name: string, value: string) => {
-        let changeTo = value;
-
-        if (name === 'collectionId') {
-            const foundCollection = credentialCollections.find(collection => collection.id === value);
-
-            if (foundCollection) changeTo = foundCollection.id;
-        }
-        setCredentialFormState(prevState => ({
-            ...prevState,
-            [name]: changeTo
-        }));
-        setFormValidationState(prevState => ({
-            ...prevState,
-            [name]: changeTo !== ''
-        }));
-    };
-
     return (
         <CustomDialog
             isOpen={open}
@@ -136,47 +94,19 @@ const CreateGeneralInfoFormDialog: FC<CreateGeneralInfoProps> = ({ onClose, open
                         <Grid item xs={12}>
                             <Typography variant="h5">{translate('Credential.GeneralInfo.Title')}</Typography>
                         </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                fullWidth
-                                label={translate('Credential.Details.Name.Label')}
-                                required
-                                name="name"
-                                InputLabelProps={{ shrink: true }}
-                                value={credentialFormState.name}
-                                onChange={handleInputChange}
-                                error={formValidationState.edited && !formValidationState.name}
-                                helperText={formValidationState.edited && !formValidationState.name && inputErrorMessages.NAME_IS_REQUIRED}
-                            ></TextField>
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TextField
-                                fullWidth
-                                label={translate('Credential.Details.Description.Label')}
-                                multiline
-                                name="description"
-                                value={credentialFormState.description}
-                                onChange={handleInputChange}
-                            ></TextField>
-                        </Grid>
-                        <Grid item xs={12} mt={2}>
-                            <CollectionDropdown
-                                disabled={!!collectionId}
-                                credentialCollections={availableCredentialCollections}
-                                selectedValue={credentialFormState.collectionId}
-                                handleChange={(event: SelectChangeEvent) => handleDropdownChange('collectionId', event.target.value)}
-                                error={formValidationState.edited && !formValidationState.collectionId}
-                                helperText={formValidationState.edited && inputErrorMessages.COLLECTION_IS_REQUIRED}
-                            />
-                        </Grid>
-                        <Grid item xs={12}>
-                            <TemplateDropdown
-                                selectedValue={credentialFormState.templateId}
-                                handleChange={(event: SelectChangeEvent) => handleDropdownChange('templateId', event.target.value)}
-                                error={formValidationState.edited && !formValidationState.templateId}
-                                helperText={formValidationState.edited && inputErrorMessages.TEMPLATE_IS_REQUIRED}
-                            />
-                        </Grid>
+                        <GeneralInfoTextFields
+                            credentialFormState={credentialFormState}
+                            setCredentialFormState={setCredentialFormState}
+                            formValidationState={formValidationState}
+                            setFormValidationState={setFormValidationState}
+                        />
+                        <GeneralInfoSelectFields
+                            collectionId={collectionId}
+                            credentialFormState={credentialFormState}
+                            setCredentialFormState={setCredentialFormState}
+                            formValidationState={formValidationState}
+                            setFormValidationState={setFormValidationState}
+                        />
                     </Grid>
                 </Form>
             </Content>
